Refactor controller modules
