+++ api-gateway/src/app.module.ts
@@ -8,6 +8,8 @@ import { JogadoresModule } from './jogadores/jogadores.module';
 import { AwsModule } from './aws/aws.module';
 import { DesafiosModule } from './desafios/desafios.module';
 import { RankingsModule } from './rankings/rankings.module';
+import { ProxyRMQModule } from './proxymq/proxymq.module';
+import { ClientProxySmartRanking } from './proxymq/client-proxy';
 
 @Module({
   imports: [
@@ -16,8 +18,10 @@ import { RankingsModule } from './rankings/rankings.module';
     DesafiosModule,
     RankingsModule,
     AwsModule,
+    ProxyRMQModule,
     ConfigModule.forRoot({ isGlobal: true }),
   ],
   controllers: [],
+  providers: [ClientProxySmartRanking],
 })
 export class AppModule {}
