+++ api-gateway/src/desafios/desafios.controller.ts
@@ -3,60 +3,33 @@
 import {
   Body,
   Controller,
-  Logger,
   Get,
   Post,
   Param,
   UsePipes,
   ValidationPipe,
-  BadRequestException,
   Put,
   Delete,
 } from '@nestjs/common';
-import { lastValueFrom } from 'rxjs';
-
-import { ClientProxySmartRanking } from '../proxymq/client-proxy';
-import { ValidacaoParametrosPipe } from 'src/common/pipes/validacao-parametros.pipe';
 
 import { CriarDesafioDTO } from './dtos/criarDesafio.dto';
 import { AtualizarDesafioDTO } from './dtos/atualizarDesafio.dto';
-import { IJogador } from '../jogadores/interfaces/jogador.interface';
-import { IDesafio } from './interfaces/desafio.interface';
-import { IDesafioStatusEnum } from './interfaces/desafio-status.enum';
 import { AtribuirDesafioPartidaDTO } from './dtos/atribuirDesafioPartida.dto';
-import { IPartida } from './interfaces/partida.interface';
+
+import { ValidacaoParametrosPipe } from 'src/common/pipes/validacao-parametros.pipe';
+import { DesafiosService } from './desafios.service';
 
 @Controller('api/v1/desafios')
 export class DesafiosController {
-  constructor(
-    private readonly clientProxySmartRaking: ClientProxySmartRanking,
-  ) {}
-
-  private logger = new Logger(DesafiosController.name);
-
-  private clientAdminBackend =
-    this.clientProxySmartRaking.getClientProxyInstance();
-
-  private clientChallenges =
-    this.clientProxySmartRaking.getClientProxyInstanceChallenges();
+  constructor(private readonly desafiosService: DesafiosService) {}
 
   @Get('/:idJogador')
   @UsePipes(ValidationPipe)
   async consultarDesafiosJogador(
     @Param('idJogador', ValidacaoParametrosPipe) idJogador: string,
   ) {
-    const jogadorCadastrado = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador', idJogador),
-    );
-
-    if (!jogadorCadastrado) {
-      throw new BadRequestException(
-        `O Jogador com id: ${idJogador}, não está cadastrado!`,
-      );
-    }
-
-    const desafiosJogador = await lastValueFrom(
-      this.clientChallenges.send('consultar-desafios-jogador', idJogador),
+    const desafiosJogador = await this.desafiosService.consultarDesafiosJogador(
+      idJogador,
     );
 
     return desafiosJogador;
@@ -65,71 +38,9 @@ export class DesafiosController {
   @Post()
   @UsePipes(ValidationPipe)
   async criarDesafio(@Body() criarDesafioDTO: CriarDesafioDTO) {
-    // Verificar se os jogadores estao cadastrados
-    const jogadoresCadastrados: IJogador[] = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogadores', ''),
-    );
+    const desafio = await this.desafiosService.criarDesafio(criarDesafioDTO);
 
-    const jogadoresNaoCadastrados = criarDesafioDTO.jogadores.filter(
-      (jogadorCadastrado) =>
-        !jogadoresCadastrados.some(
-          (jogador) => jogador._id === jogadorCadastrado._id,
-        ),
-    );
-
-    if (jogadoresNaoCadastrados.length > 0) {
-      throw new BadRequestException(
-        `Jogadore(s) ${jogadoresNaoCadastrados
-          .map((item) => item._id)
-          .join(', ')} não cadastrado(s)!`,
-      );
-    }
-
-    // Verificar se a categoria esta cadastrada
-    await lastValueFrom(
-      this.clientAdminBackend.send(
-        'consultar-categoria',
-        criarDesafioDTO.categoria,
-      ),
-    );
-
-    // Verificando se o solicitante é um jogador da partida
-    const solicitanteFazParteDesafio = criarDesafioDTO.jogadores.find(
-      (jogador) => jogador._id === criarDesafioDTO.solicitante,
-    );
-
-    if (!solicitanteFazParteDesafio) {
-      throw new BadRequestException(
-        `O Solicitante com id ${criarDesafioDTO.solicitante}, não faz parte do desafio! Informe um ID solicitante que faça parte do desafio!`,
-      );
-    }
-
-    // Verificando se os jogadores realmente pertencem a categoria que foi informada no body da requisição
-    const jogadorCategoriaNaoCorrespondente = criarDesafioDTO.jogadores.filter(
-      (jogador) =>
-        jogadoresCadastrados.some(
-          (jogadorCadastrado) =>
-            jogador._id == jogadorCadastrado._id &&
-            jogadorCadastrado.categoria !== criarDesafioDTO.categoria,
-        ),
-    );
-
-    if (jogadorCategoriaNaoCorrespondente.length > 0) {
-      throw new BadRequestException(
-        `Jogadore(s) com id: (${jogadorCategoriaNaoCorrespondente
-          .map((item) => item._id)
-          .join(', ')}), não está com a categoria correspondente!`,
-      );
-    }
-
-    const dataCriarDesafio = {
-      ...criarDesafioDTO,
-      status: IDesafioStatusEnum.PENDENTE,
-    };
-
-    return await lastValueFrom(
-      this.clientChallenges.emit('criar-desafio', dataCriarDesafio),
-    );
+    return desafio;
   }
 
   @Post('/:idDesafio/partida')
@@ -138,49 +49,12 @@ export class DesafiosController {
     @Param('idDesafio', ValidacaoParametrosPipe) idDesafio: string,
     @Body() atribuirDesafioPartidaDTO: AtribuirDesafioPartidaDTO,
   ) {
-    const desafioCadastrado: IDesafio = await lastValueFrom(
-      this.clientChallenges.send('consultar-desafio', idDesafio),
+    const partidaAtribuida = await this.desafiosService.atribuirDesafioPartida(
+      idDesafio,
+      atribuirDesafioPartidaDTO,
     );
 
-    if (!desafioCadastrado) {
-      throw new BadRequestException(
-        `Desafio com id ${idDesafio}, não está cadastrado!`,
-      );
-    }
-
-    if (desafioCadastrado.status === IDesafioStatusEnum.REALIZADO) {
-      throw new BadRequestException(
-        `Desafio com status REALIZADO não pode ser atribuido a nenhuma partida!`,
-      );
-    }
-
-    if (desafioCadastrado.status !== IDesafioStatusEnum.ACEITO) {
-      throw new BadRequestException(
-        `Desafio informado contem status invalido!! Apenas desafios com status ACEITO podem ser atribuidos a uma partida!`,
-      );
-    }
-
-    if (!desafioCadastrado.jogadores.includes(atribuirDesafioPartidaDTO.def)) {
-      throw new BadRequestException(
-        `Jogador vencedor com id ${atribuirDesafioPartidaDTO.def}, não faz parte do desafio!`,
-      );
-    }
-
-    const novaPartida: IPartida = {
-      categoria: desafioCadastrado.categoria,
-      def: atribuirDesafioPartidaDTO.def,
-      resultado: atribuirDesafioPartidaDTO.resultado,
-      jogadores: desafioCadastrado.jogadores,
-      desafio: desafioCadastrado._id,
-    };
-
-    /* const partidaCriada = await lastValueFrom(
-      this.clientChallenges.emit('criar-partida', novaPartida),
-    ); */
-
-    return this.clientChallenges.emit('criar-partida', novaPartida);
-
-    // return partidaCriada;
+    return partidaAtribuida;
   }
 
   @Put('/:idDesafio')
@@ -189,32 +63,9 @@ export class DesafiosController {
     @Param('idDesafio', ValidacaoParametrosPipe) idDesafio: string,
     @Body() atualizarDesafioDTO: AtualizarDesafioDTO,
   ) {
-    const desafioCadastrado = await lastValueFrom(
-      this.clientChallenges.send('consultar-desafio', idDesafio),
-    );
-
-    if (!desafioCadastrado) {
-      throw new BadRequestException(
-        `Desafio com id: ${idDesafio}, não esta cadastrado!`,
-      );
-    }
-
-    if (desafioCadastrado.status !== IDesafioStatusEnum.PENDENTE) {
-      throw new BadRequestException(
-        `O Desafio está com status invalido! Status Pendente apenas pode ser atualizado!!`,
-      );
-    }
-
-    const desafioAtualizar = {
-      desafio: {
-        ...desafioCadastrado,
-        status: atualizarDesafioDTO.status,
-      },
+    const desafioAtualizado = await this.atualizarDesafio(
       idDesafio,
-    };
-
-    const desafioAtualizado = await lastValueFrom(
-      this.clientChallenges.emit('atualizar-desafio', desafioAtualizar),
+      atualizarDesafioDTO,
     );
 
     return desafioAtualizado;
@@ -225,18 +76,6 @@ export class DesafiosController {
   async deletarDesafio(
     @Param('idDesafio', ValidacaoParametrosPipe) idDesafio: string,
   ) {
-    const desafioCadastrado = await lastValueFrom(
-      this.clientChallenges.send('consultar-desafio', idDesafio),
-    );
-
-    if (!desafioCadastrado) {
-      throw new BadRequestException(
-        `Desafio com id: ${idDesafio} não esta cadastrado!`,
-      );
-    }
-
-    await lastValueFrom(
-      this.clientChallenges.emit('deletar-desafio', idDesafio),
-    );
+    await this.desafiosService.deletarDesafio(idDesafio);
   }
 }
