+++ api-gateway/src/jogadores/jogadores.controller.ts
@@ -2,16 +2,13 @@
 
 import {
   Controller,
-  Logger,
   Post,
   ValidationPipe,
   Body,
   Get,
   Put,
   Param,
-  BadRequestException,
 } from '@nestjs/common';
-import { lastValueFrom } from 'rxjs';
 import {
   Delete,
   UploadedFile,
@@ -20,110 +17,39 @@ import {
 } from '@nestjs/common/decorators';
 import { FileInterceptor } from '@nestjs/platform-express';
 
-import { ClientProxySmartRanking } from '../proxymq/client-proxy';
-
-import { ValidacaoParametrosPipe } from 'src/common/pipes/validacao-parametros.pipe';
-
 import { CriarJogadorDTO } from './dtos/criarJogador.dto';
 import { AtualizarJogadorDTO } from './dtos/atualizarjogador.dto';
-import { AwsService } from '../aws/aws.service';
+
+import { ValidacaoParametrosPipe } from '../common/pipes/validacao-parametros.pipe';
+import { JogadoresService } from './jogadores.service';
 
 @Controller('api/v1/jogadores')
 export class JogadoresController {
-  constructor(
-    private readonly clientProxySmartRaking: ClientProxySmartRanking,
-    private readonly awsService: AwsService,
-  ) {}
-
-  private logger = new Logger(JogadoresController.name);
-
-  private clientAdminBackend =
-    this.clientProxySmartRaking.getClientProxyInstance();
+  constructor(private readonly jogadoresService: JogadoresService) {}
 
   @Post()
   @UsePipes(ValidationPipe)
   async criarJogador(@Body() criarJogadorDTO: CriarJogadorDTO) {
-    const jogador = {
-      nome: criarJogadorDTO.nome,
-      email: criarJogadorDTO.email,
-      telefoneCelular: criarJogadorDTO.telefoneCelular,
-    };
-
-    const jogadorExistente = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador-email', jogador.email),
-    );
-
-    if (jogadorExistente) {
-      throw new BadRequestException(
-        `Jogador com email ${jogador.email} já cadastrado!`,
-      );
-    }
-
-    const categoria = criarJogadorDTO.categoria;
-
-    const categoriaExistente = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-categoria', categoria),
-    );
-
-    if (Object.values(categoriaExistente).length > 0) {
-      this.clientAdminBackend.emit('criar-jogador', {
-        categoria,
-        jogador,
-      });
-    } else {
-      throw new BadRequestException('Categoria não cadastrada!');
-    }
+    await this.jogadoresService.criarJogador(criarJogadorDTO);
   }
 
   @Post('/:id/upload')
   @UseInterceptors(FileInterceptor('file'))
   async uploadArquivo(@UploadedFile() file, @Param('id') id: string) {
-    this.logger.log(`file ${file}`);
+    const uploadArquivo = await this.jogadoresService.uploadArquivo(file, id);
 
-    // Verificar se o jogador esta cadastrado
-    const jogadorExiste = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador', id),
-    );
-
-    if (!jogadorExiste) {
-      throw new BadRequestException(`Jogador não encontrado!`);
-    }
-
-    // Enviar o arquivo para S3
-    // Recuperar a URL de acesso
-    const { url: urlFotoJogador } = await this.awsService.uploadArquivo(
-      file,
-      id,
-    );
-
-    // Atualizar o atributo URL da entidade de jogador
-    const atualizarJogadorDTO: AtualizarJogadorDTO = {};
-    atualizarJogadorDTO.urlFotoJogador = urlFotoJogador;
-
-    this.clientAdminBackend.emit('atualizar-jogador', {
-      id,
-      jogador: atualizarJogadorDTO,
-    });
-
-    // Retornar o jogador atualizado para o cliente
-    return await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador', id),
-    );
+    return uploadArquivo;
   }
 
   @Get()
   async consultarJogadores() {
-    return await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogadores', ''),
-    );
+    return await this.jogadoresService.consultarJogadores();
   }
 
   @Get('/:id')
   @UsePipes(ValidationPipe)
   async consultarJogador(@Param('id', ValidacaoParametrosPipe) id: string) {
-    return await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador', id),
-    );
+    return await this.jogadoresService.consultarJogador(id);
   }
 
   @Put('/:idJogador')
@@ -132,32 +58,14 @@ export class JogadoresController {
     @Body() atualizarJogadorDTO: AtualizarJogadorDTO,
     @Param('idJogador', ValidacaoParametrosPipe) idJogador: string,
   ) {
-    const jogadorExistente = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador', idJogador),
+    await this.jogadoresService.atualizarJogador(
+      atualizarJogadorDTO,
+      idJogador,
     );
-
-    if (!jogadorExistente) {
-      throw new BadRequestException(
-        `Jogador com id: ${idJogador}, não encontrado!`,
-      );
-    }
-
-    this.clientAdminBackend.emit('atualizar-jogador', {
-      id: idJogador,
-      jogador: atualizarJogadorDTO,
-    });
   }
 
   @Delete('/:_id')
   async deletarJogador(@Param('_id', ValidacaoParametrosPipe) _id: string) {
-    const jogadorExistente = await lastValueFrom(
-      this.clientAdminBackend.send('consultar-jogador', _id),
-    );
-
-    if (!jogadorExistente) {
-      throw new BadRequestException(`Jogador com id: ${_id}, não encontrado!`);
-    }
-
-    this.clientAdminBackend.emit('deletar-jogador', _id);
+    await this.jogadoresService.deletarJogador(_id);
   }
 }
