+++ api-gateway/src/rankings/rankings.controller.ts
@@ -1,31 +1,23 @@
 /* eslint-disable prettier/prettier */
 
-import { BadRequestException, Controller, Get, Query } from '@nestjs/common';
-import { lastValueFrom } from 'rxjs';
+import { Controller, Get, Query } from '@nestjs/common';
 
-import { ClientProxySmartRanking } from '../proxymq/client-proxy';
+import { RankingsService } from './rankings.service';
 
 @Controller('api/v1/rankings')
 export class RankingsController {
-  constructor(private clientProxySmartRanking: ClientProxySmartRanking) {}
-
-  private clientRankingsBackEnd =
-    this.clientProxySmartRanking.getClientProxyInstanceRankings();
+  constructor(private readonly rankingsService: RankingsService) {}
 
   @Get()
   async consultarRankings(
     @Query('idCategoria') idCategoria: string,
     @Query('dataRef') dataRef: string,
   ) {
-    if (!idCategoria) {
-      throw new BadRequestException('O Id da Categoria não foi informado!');
-    }
-
-    return await lastValueFrom(
-      this.clientRankingsBackEnd.send('consultar-rankings', {
-        idCategoria,
-        dataRef: dataRef ? dataRef : '',
-      }),
+    const rankings = await this.rankingsService.consultarRankings(
+      idCategoria,
+      dataRef,
     );
+
+    return rankings;
   }
 }
